Ensure AutoMod has a chance to run
Reddit's shoddy implementation of filtering means that if you browse a subreddit's 'new' feed, you will see new posts as-they're-made, *before* AutoModerator runs (even if AutoModerator filters all content for review, i.e. r/netsec). As such, a scam was posted to the r/netsec deeplink bot, even though this scam was only visible for 0-3min to Reddit-native r/netsec viewers. A delay of 15 minutes has been added before posts are eligible to be crossposted, and this may be made configurable for additional usecases such as subreddits which may be hand-moderated daily.